Move interpret_identifier to its own file
This function is needed in convert_expression_string_to_predicate, so it needs to be in another file to avoid a circular import.
